metodos de pago

+++ app/Http/Livewire/CategoriaController.php
@@ -5,12 +5,13 @@ namespace App\Http\Livewire;
 use Livewire\Component;
 use App\Models\Auditoria;
 use App\Models\Categoria;
+use App\Models\Producto;
 use DB;
 
 class CategoriaController extends Component
 {
 	public  $descripcion, $margen, $descripcion_soft_deleted, $id_soft_deleted;
-    public  $selected_id, $search, $comentario = '';  
+    public  $selected_id, $search;  
     public $comercioId, $action = 1, $recuperar_registro = 0;
 
     public function render()
@@ -33,23 +34,22 @@ class CategoriaController extends Component
                         ->where('comercio_id', $this->comercioId)->get()
         ]);
        }
-    }
-
+    }    
+    protected $listeners = [
+        'deleteRow'=>'destroy'        
+    ];
     public function doAction($action)
     {
         $this->action = $action;
         $this->resetInput();
     }
-
     private function resetInput()
     {
         $this->descripcion = '';
         $this->margen = '';
         $this->selected_id = null;    
         $this->search = '';
-        $this->comentario = '';
     }
-
     public function edit($id)
     {
         $this->action = 2;
@@ -58,19 +58,25 @@ class CategoriaController extends Component
         $this->descripcion = $record->descripcion;
         $this->margen = $record->margen;
     }
-
     public function volver()
     {
         $this->recuperar_registro = 0;
         $this->resetInput();
         return; 
     }
-
     public function RecuperarRegistro($id)
     {
         DB::begintransaction();
         try{
             Categoria::onlyTrashed()->find($id)->restore();
+            $audit = Auditoria::create([
+                'item_deleted_id' => $id,
+                'tabla'           => 'Categorías',
+                'estado'          => '1',
+                'user_delete_id'  => auth()->user()->id,
+                'comentario'      => '',
+                'comercio_id'     => $this->comercioId
+            ]);
             session()->flash('msg-ok', 'Registro recuperado');
             $this->volver();
             
@@ -80,7 +86,6 @@ class CategoriaController extends Component
             session()->flash('msg-error', '¡¡¡ATENCIÓN!!! El registro no se recuperó...');
         }
     }
-
     public function StoreOrUpdate()
     { 
         $this->validate([
@@ -93,15 +98,14 @@ class CategoriaController extends Component
                 ->where('id', '<>', $this->selected_id)
                 ->where('comercio_id', $this->comercioId)
                 ->withTrashed()->get();
-
-                if($existe->count() > 0 && $existe[0]->deleted_at != null) {
-                    session()->flash('info', 'La Categoría que desea modificar ya existe pero fué eliminada anteriormente, para recuperarla haga click en el botón "Recuperar registro"');
+                if($existe->count() && $existe[0]->deleted_at != null) {
+                    //session()->flash('info', 'La Categoría que desea modificar ya existe pero fué eliminada anteriormente, para recuperarla haga click en el botón "Recuperar registro"');
                     $this->action = 1;
                     $this->recuperar_registro = 1;
                     $this->descripcion_soft_deleted = $existe[0]->descripcion;
                     $this->id_soft_deleted = $existe[0]->id;
                     return;
-                }elseif( $existe->count() > 0) {
+                }elseif($existe->count()) {
                     session()->flash('info', 'La Categoría ya existe...');
                     $this->resetInput();
                     return;
@@ -110,14 +114,14 @@ class CategoriaController extends Component
                 $existe = Categoria::where('descripcion', $this->descripcion)
                 ->where('comercio_id', $this->comercioId)->withTrashed()->get();
 
-                if($existe->count() > 0 && $existe[0]->deleted_at != null) {
-                    session()->flash('info', 'La Categoría que desea agregar ya existe pero fué eliminado anteriormente, para recuperarlo haga click en el botón "Recuperar registro"');
+                if($existe->count() && $existe[0]->deleted_at != null) {
+                    //session()->flash('info', 'La Categoría que desea agregar ya existe pero fué eliminado anteriormente, para recuperarlo haga click en el botón "Recuperar registro"');
                     $this->action = 1;
                     $this->recuperar_registro = 1;
                     $this->descripcion_soft_deleted = $existe[0]->descripcion;
                     $this->id_soft_deleted = $existe[0]->id;
                     return;
-                }elseif($existe->count() > 0 ) {
+                }elseif($existe->count()) {
                     session()->flash('info', 'La Categoría ya existe...');
                     $this->resetInput();
                     return;
@@ -147,31 +151,30 @@ class CategoriaController extends Component
         }
         $this->resetInput();
         return;
-    }
-    
-    protected $listeners = [
-        'deleteRow'=>'destroy'        
-    ];  
-       
-    public function destroy($id)
+    }        
+    public function destroy($id, $comentario)
     {
         if ($id) {
-            DB::begintransaction();
-            try{
-                $categoria = Categoria::find($id)->delete();
-                $audit = Auditoria::create([
-                    'item_deleted_id' => $id,
-                    'tabla' => 'Categorias',
-                    'user_delete_id' => auth()->user()->id,
-                    'comentario' => $this->comentario,
-                    'comercio_id' => $this->comercioId
-                ]);
-                session()->flash('msg-ok', 'Registro eliminado con éxito!!');
-                DB::commit();               
-            }catch (Exception $e){
-                DB::rollback();
-                session()->flash('msg-error', '¡¡¡ATENCIÓN!!! El registro no se eliminó...');
-            }
+            $record = Producto::where('categoria_id', $id)->get();
+            if(!$record->count()){
+                DB::begintransaction();
+                try{
+                    $categoria = Categoria::find($id)->delete();
+                    $audit = Auditoria::create([
+                        'item_deleted_id' => $id,
+                        'tabla' => 'Categorías',
+                        'user_delete_id' => auth()->user()->id,
+                        'comentario' => $comentario,
+                        'comercio_id' => $this->comercioId
+                    ]);
+                    DB::commit();  
+                    $this->emit('registroEliminado');             
+                }catch (Exception $e){
+                    DB::rollback();
+                    session()->flash('msg-error', '¡¡¡ATENCIÓN!!! El registro no se eliminó...');
+                }    
+            }else $this->emit('eliminarRegistro');
+         
             $this->resetInput();
             return;
         }
